+++ EssentialApp/EssentialAppTests/FeedImageDataLoaderCacheDecoratorTests.swift
@@ -18,7 +18,7 @@ class FeedImageDataLoaderCacheDecorator: FeedImageDataLoader {
 	}
 }
 
-class FeedImageDataLoaderCacheDecoratorTests: XCTestCase {
+class FeedImageDataLoaderCacheDecoratorTests: XCTestCase, FeedImageDataLoaderTestCase {
 	
 	func test_init_doesNotLoadImageData() {
 		let (_, loader) = makeSUT()
@@ -71,28 +71,5 @@ class FeedImageDataLoaderCacheDecoratorTests: XCTestCase {
 		trackForMemoryLeaks(sut, file: file, line: line)
 		return (sut, loader)
 	}
-	
-	private func expect(_ sut: FeedImageDataLoader, toCompleteWith expectedResult: FeedImageDataLoader.Result, when action: () -> Void, file: StaticString = #file, line: UInt = #line) {
-		let exp = expectation(description: "Wait for load completion")
 		
-		_ = sut.loadImageData(from: anyURL()) { receivedResult in
-			switch (receivedResult, expectedResult) {
-			case let (.success(receivedFeed), .success(expectedFeed)):
-				XCTAssertEqual(receivedFeed, expectedFeed, file: file, line: line)
-				
-			case (.failure, .failure):
-				break
-				
-			default:
-				XCTFail("Expected \(expectedResult), got \(receivedResult) instead", file: file, line: line)
-			}
-			
-			exp.fulfill()
-		}
-		
-		action()
-		
-		wait(for: [exp], timeout: 1.0)
-	}
-	
 }
