+++ EssentialApp/EssentialAppTests/FeedImageDataLoaderWithFallbackCompositeTests.swift
@@ -6,7 +6,7 @@ import XCTest
 import EssentialFeed
 import EssentialApp
 
-class FeedImageDataLoaderWithFallbackCompositeTests: XCTestCase {
+class FeedImageDataLoaderWithFallbackCompositeTests: XCTestCase, FeedImageDataLoaderTestCase {
 	
 	func test_init_doesNotLoadImageData() {
 		let (_, primaryLoader, fallbackLoader) = makeSUT()
@@ -98,29 +98,6 @@ class FeedImageDataLoaderWithFallbackCompositeTests: XCTestCase {
 		trackForMemoryLeaks(fallbackLoader, file: file, line: line)
 		trackForMemoryLeaks(sut, file: file, line: line)
 		return (sut, primaryLoader, fallbackLoader)
-	}
-	
-	private func expect(_ sut: FeedImageDataLoader, toCompleteWith expectedResult: FeedImageDataLoader.Result, when action: () -> Void, file: StaticString = #file, line: UInt = #line) {
-		let exp = expectation(description: "Wait for load completion")
-		
-		_ = sut.loadImageData(from: anyURL()) { receivedResult in
-			switch (receivedResult, expectedResult) {
-			case let (.success(receivedFeed), .success(expectedFeed)):
-				XCTAssertEqual(receivedFeed, expectedFeed, file: file, line: line)
-				
-			case (.failure, .failure):
-				break
-				
-			default:
-				XCTFail("Expected \(expectedResult), got \(receivedResult) instead", file: file, line: line)
-			}
-			
-			exp.fulfill()
-		}
-		
-		action()
-		
-		wait(for: [exp], timeout: 1.0)
 	}
 		
 }
